+++ clients/bendini/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `register` command takes publisher name and email if not provided they will be
+  retrieved from the auth service.
+
 ## [1.0.0] - 2021-07-03
 
 ### Added