+++ services/avery/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Implemented endpoint for GetIdentity
+
 ## [1.2.1] - 2021-10-21
 
 ### Fixed