+++ services/quinn/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Quinn stores publisher, with name and email
+
 ## [1.0.0] - 2021-07-03
 
 ### Added